py/requirements.txt
requests
html5lib
bs4
mysql-connector